Improve recording time label update mechanism
Provide a dedicated structure that keep track of the time displayed on
the label and call the GUI function to display it only if it has
changed. Also fix the time tracking of recording: if a second file is
recording while acquisition is still on, the time tracked will be reset.

Change-Id: I2779522bea65f16e15ded4fba2fbc5f71afecd96