Set store type when creating xdf channel
When creating and configure XDF file, the stored type was not set and
left to the default one of the data file type used for writing. With
BDF and EDF there was no issue since only one type was allowed
(respectively int24 and int16). However with GDF which support different
stored type, all channels were recorded as float (even integer
channels).

Now when configuring a XDF channel, we aim at using type int32 if
integral channel as reported by eegdev, and float otherwise. It is only
a targetted type as we will store in the closest supported type of the
data file.

Change-Id: I63faacb791f7502cc7267de8a249e398f1d5da68